refactor: make enactToggleSignMode be module
