+++ public/index.html
@@ -82,7 +82,6 @@
 
   <script src="js/globals.js"></script>
   <script type="module" src="js/app.js"></script>
-  <script src="js/utils/toggle-sign-mode.js"></script>
   <script src="js/utils/show-error.js"></script>
   <script src="js/utils/set-waiting-state.js"></script>
 </body>
